Move PromptContext to its own module
